Pop preferentially from LIFO stack upon commit
IfStack can push particles into two different stacks, but always pops
from the first if it is nonempty. In the leading muon bias application,
this causes the entire primary stack to the processed and shifted to the
secondary stack when the shower switches state from PENDING to
COMMITTED, which above a few PeV can create more particles than the
(fixed-size) stack can store, which in turn causes CORSIKA to silently
drop all but the leading sub-shower. To fix this, extend IfStack to take
an optional callback to determine which stack to pop from first. This
defaults to the first stack as before.